test: split remaining mixed test files into unit and integration
