+++ src/chat-slash.test.ts
@@ -1,7 +1,6 @@
-import { afterEach, beforeEach, describe, expect, test } from "bun:test";
+import { beforeEach, describe, expect, test } from "bun:test";
 import { isKnownSlashToken, shouldAutocompleteSlashSubmit, slashCommandHelp, suggestSlashCommands } from "./chat-slash";
-import { loadSkills, resetSkillCache } from "./skills";
-import { tempDir, writeSkill } from "./test-utils";
+import { resetSkillCache } from "./skills";
 
 beforeEach(() => resetSkillCache());
 
@@ -86,31 +85,5 @@ describe("chat-slash helpers", () => {
     expect(slashCommandHelp("/model")).toBe("change model");
     expect(slashCommandHelp("/model work")).toBe("");
     expect(slashCommandHelp("/unknown")).toBe("");
-  });
-
-  describe("with loaded skills", () => {
-    const { createDir, cleanupDirs } = tempDir();
-    afterEach(() => {
-      resetSkillCache();
-      cleanupDirs();
-    });
-
-    test("skill names appear in suggestions", async () => {
-      const tmpDir = createDir("acolyte-slash-skill-");
-      writeSkill(tmpDir, "dogfood", "---\nname: dogfood\ndescription: Test\n---", "# Test");
-      await loadSkills(tmpDir);
-
-      const suggestions = suggestSlashCommands("/dog");
-      expect(suggestions).toContain("/dogfood");
-    });
-
-    test("isKnownSlashToken recognizes skill names", async () => {
-      const tmpDir = createDir("acolyte-slash-known-");
-      writeSkill(tmpDir, "dogfood", "---\nname: dogfood\ndescription: Test\n---", "# Test");
-      await loadSkills(tmpDir);
-
-      expect(isKnownSlashToken("/dogfood")).toBe(true);
-      expect(isKnownSlashToken("/nonexistent")).toBe(false);
-    });
   });
 });
