+++ src/cli-trace.int.test.ts
