+++ src/lifecycle-effects.test.ts
@@ -1,8 +1,6 @@
-import { afterEach, describe, expect, test } from "bun:test";
-import { mkdirSync } from "node:fs";
-import { join } from "node:path";
+import { describe, expect, test } from "bun:test";
 import { formatEffect, installEffect, lintEffect } from "./lifecycle-effects";
-import { createRunContext, tempDir } from "./test-utils";
+import { createRunContext } from "./test-utils";
 
 function ctxWith(overrides: Parameters<typeof createRunContext>[0] = {}) {
   return createRunContext({
@@ -35,9 +33,6 @@ describe("formatEffect", () => {
   });
 });
 
-const { createDir, cleanupDirs } = tempDir();
-afterEach(() => cleanupDirs());
-
 describe("installEffect", () => {
   test("returns done when workspace is undefined", () => {
     const ctx = ctxWith({ workspace: undefined });
@@ -48,17 +43,6 @@ describe("installEffect", () => {
     const ctx = ctxWith({
       policy: { ...createRunContext().policy, installCommand: undefined },
     });
-    expect(installEffect.run(ctx, [])).toEqual({ type: "done" });
-  });
-
-  test("skips install when depsDir exists", () => {
-    const ws = createDir("acolyte-install-");
-    mkdirSync(join(ws, "node_modules"), { recursive: true });
-    const ctx = ctxWith({
-      workspace: ws,
-      policy: { ...createRunContext().policy, installCommand: { bin: "npm", args: ["install"] } },
-    });
-    ctx.session.workspaceProfile = { depsDir: "node_modules" };
     expect(installEffect.run(ctx, [])).toEqual({ type: "done" });
   });
 });
