+++ src/server-chat-runtime.test.ts
@@ -1,13 +1,7 @@
-import { afterEach, describe, expect, test } from "bun:test";
-import { join } from "node:path";
+import { describe, expect, test } from "bun:test";
 import { appConfig } from "./app-config";
 import { logLifecycleDebugEntry, runChatRequest } from "./server-chat-runtime";
 import type { StreamErrorPayload } from "./server-contract";
-import { tempDir } from "./test-utils";
-import { createTraceStore } from "./trace-store";
-
-const { createDir, cleanupDirs } = tempDir();
-afterEach(cleanupDirs);
 
 describe("server chat runtime", () => {
   test("logLifecycleDebugEntry logs agent debug entry", () => {
@@ -30,27 +24,6 @@ describe("server chat runtime", () => {
     expect(logs).toHaveLength(1);
     expect(logs[0]?.message).toBe("agent debug");
     expect(logs[0]?.fields?.event).toBe("lifecycle.memory.commit_done");
-  });
-
-  test("logLifecycleDebugEntry dual-writes to trace store", () => {
-    const dir = createDir("acolyte-trace-srv-");
-    const store = createTraceStore(join(dir, "trace.db"));
-    logLifecycleDebugEntry({
-      requestId: "err_xyz789",
-      taskId: "task_dual",
-      sessionId: "sess_dual",
-      event: "lifecycle.start",
-      sequence: 1,
-      eventTs: "2026-03-20T12:00:00.000Z",
-      fields: { model: "gpt-5", mode: "work" },
-      logInfo: () => {},
-      traceStore: store,
-    });
-    const lines = store.listByTaskId("task_dual");
-    expect(lines).toHaveLength(1);
-    expect(lines[0]?.fields.event).toBe("lifecycle.start");
-    expect(lines[0]?.fields.model).toBe("gpt-5");
-    store.close();
   });
 
   test("vercel provider error mentions AI_GATEWAY_API_KEY", async () => {
